refs qorelanguage/qore#4091 major progress on python imports, function imports, and dynamic class creation in Java for external Qore/Python values

+++ test/java/src/org/qore/jni/test/ClosureTest3.java
@@ -3,6 +3,6 @@ package org.qore.jni.test;
 import org.qore.jni.*;
 
 public class ClosureTest3 implements QoreClosureMarker {
-    public void call2(int i) {
+    public void call2(long i) {
     }
 }
